cleanup, moving packages to nargo workspace, documenting
