+++ recursion/package.json
@@ -9,8 +9,6 @@
     "test": "hardhat test"
   },
   "dependencies": {
-    "@aztec/bb.js": "0.7.2",
-    "@noir-lang/acvm_js": "0.26.1",
     "@noir-lang/backend_barretenberg": "^0.7.10",
     "@noir-lang/noir_js": "^0.16.0",
     "@nomicfoundation/hardhat-chai-matchers": "^2.0.0",
@@ -22,10 +20,8 @@
     "@typechain/ethers-v6": "^0.4.0",
     "@typechain/hardhat": "^8.0.0",
     "@types/chai": "^4.2.0",
-    "axios": "^1.3.4",
     "chai": "^4.2.0",
     "dotenv": "^16.0.3",
-    "fflate": "^0.8.0",
     "hardhat": "^2.18.0",
     "hardhat-gas-reporter": "^1.0.8",
     "next": "latest",
